Add files via upload
In this lab we used a puredata patch to read gyroscope information from an app downloaded on the mobile phone. Inorder to downlad and run, Wekinator, Puredata Extended, and an appropriate mobile app relative to your operating system must be downloaded. The mobile phone and computer running the patches must also be connected to the same server.