Add files via upload
The goal of Lab 1 was to map audio captured with the application Various_Audio_inputs and map them to a visual output using a puredata patch. My idea was to train the ML models from wekinator to recognize different ukulele chords, and display the appropriate chord chart when a particular chord was strummed. The intended use would be for beginner students in music lesson settings to facilitate playing along with the teacher. In order to download and run the patch, Wekinator, Puredata Extended, and Various_Audio_Inputs must be downloaded.